# Improve list optimisation in `CountDown`