@viceroypenguin 2023-01-23
Two things:
1. I would rerun the performance benchmark after making the changes to fix the bugs above. The missed `count` value could have a performance impact on the behavior.
2. If possible, I would copy the before/after into LinqPAD for benchmarking. The reason is because two separate benchmarks, even on the same machine, are not always comparable (other processes, etc.). It is more accurate to have a `Before()` method and an `After()` method in LinqPAD and benchmark them against each other in the same run.