@atifaziz 2023-01-23
> 1. I would rerun the performance benchmark after making the changes to fix the bugs above. The missed `count` value could have a performance impact on the behavior.

45% improvement:

|            Method |               Source |     Mean |    Error |   StdDev |   Median | Allocated |
|------------------ |--------------------- |---------:|---------:|---------:|---------:|----------:|
| MoreLinqCountDown |       Int32[1000000] | 11.27 ms | 0.316 ms | 0.923 ms | 11.02 ms |     100 B |
| MoreLinqCountDown | Syste(...)rator [36] | 20.61 ms | 0.707 ms | 2.041 ms | 20.58 ms |     179 B |


> 2. If possible, I would copy the before/after into LinqPAD for benchmarking. The reason is because two separate benchmarks, even on the same machine, are not always comparable (other processes, etc.). It is more accurate to have a `Before()` method and an `After()` method in LinqPAD and benchmark them against each other in the same run.

Yes, I am aware and admit taking a shortcut there. The difference was large enough and comparable in the many runs I did. What's more important is the comparison between list and sequence processing in the same ran than the before and after.